+++ tests/SimConnect.NET.Tests.Net8/Tests/TestRunner.cs
@@ -31,6 +31,8 @@ namespace SimConnect.NET.Tests.Net8.Tests
                 new InputEventTests(),
                 new InputEventValueTests(),
                 new PerformanceTests(),
+                new SystemEventSubscriptionTests(),
+                new SystemEventStateTests(),
             };
         }
 
@@ -142,6 +144,9 @@ namespace SimConnect.NET.Tests.Net8.Tests
                         break;
                     case "--verbose":
                         options.Verbose = true;
+                        break;
+                    case "--test-events":
+                        options.Categories.Add("System Event");
                         break;
                 }
             }